fix sim time

+++ app/ssaparamscan/ssaparamscantask.hh
@@ -99,7 +99,7 @@ protected:
   /** Holds the current iteration number. */
   size_t _current_iteration;
   /** Holds the simulation time. */
-  size_t _simulation_time;
+  double _simulation_time;
 
 public:
   explicit SSAParamScanTask(const Config &config, QObject *parent=0);
